visualizations
This file contains visualizations of a covid-19 vaccination supply statewide in India. These analyses can be done, with help of Microsoft Power BI. I do not have the business email id or email id having the extension of the company name. so that I was not able to create the dashboard.